fix: added missing check

+++ phpmyfaq/src/phpMyFAQ/Controller/Administration/Api/InstanceController.php
@@ -72,15 +72,19 @@ final class InstanceController extends AbstractController
             return $this->json(['error' => 'Cannot create instance: wrong URL'], Response::HTTP_BAD_REQUEST);
         }
 
+        $faqInstanceClient = new Client($configuration);
+        $faqInstanceClient->setFileSystem(new Filesystem());
+        if (!$faqInstanceClient->isValidClientUrl($url)) {
+            return $this->json(['error' => 'Cannot create instance: wrong URL'], Response::HTTP_BAD_REQUEST);
+        }
+
         $data = new InstanceEntity();
         $data->setUrl($url)->setInstance($instance)->setComment($comment);
 
         $faqInstance = $this->container->get(id: 'phpmyfaq.instance');
         $instanceId = $faqInstance->create($data);
 
-        $faqInstanceClient = new Client($configuration);
         $faqInstanceClient->createClient($faqInstance);
-        $faqInstanceClient->setFileSystem(new Filesystem());
 
         $urlParts = parse_url($data->getUrl());
         $hostname = $urlParts['host'];
@@ -165,6 +169,10 @@ final class InstanceController extends AbstractController
             $client = new Client($configuration);
             $client->setFileSystem(new Filesystem());
             $clientData = $client->getById($instanceId);
+            if (!$client->isValidClientUrl((string) $clientData->url)) {
+                return $this->json(['error' => $instanceId], Response::HTTP_BAD_REQUEST);
+            }
+
             if (1 !== $instanceId && $client->deleteClientFolder($clientData->url) && $client->delete($instanceId)) {
                 return $this->json(['deleted' => $instanceId], Response::HTTP_OK);
             }
